- use a 16bit value for fnstsw as required by newer binutils.

+++ libm/i386/ftestexcept.c
@@ -24,7 +24,7 @@
 int
 fetestexcept (int excepts)
 {
-  int temp;
+  short temp;
   int xtemp = 0;
 
   /* Get current exceptions.  */
